llamaindex-server/pyproject.toml
[project]
name = "llamaindex-server"
version = "0.1.0"
description = "LlamaIndex RAG server for Slack AI Assistant"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "llama-index-core>=0.10.0",
    "llama-index-llms-gemini>=0.2.0",
    "llama-index-embeddings-google-genai>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]

# Don't try to build as a package - this is just an application
package = false
